Created services and routes
- moved all routing from app.py to routes/ and moved db connection to services
- blueprints handle their own routing now
- app.py will only handle initialization of flask app and registering the blue prints
