+++ static/js/index.js
@@ -13,7 +13,7 @@
             apiToken: apiToken 
         };
 
-        const response = await fetch('http://127.0.0.1:5000/verify-user', {
+        const response = await fetch('http://127.0.0.1:5000/', {
             
             method: 'POST',
